src/requirements.txt
celery
django>=2.0
django-allauth
django-channels
django-invitations
django-model-utils
django-qr-code
django-redis
djangorestframework
django-rest-auth
django_smtp_ssl
django-money
eth-abi<2
ethtoken
factory_boy
gnosis-py
psycopg2-binary
web3>=4,<5
